Polymap desugaring cleanup
- Renamed MapDesugarHelper to PolyMapDesugarHelper.
- Heap module now exposes a field type constructor instead of a PolyMapDesugarHelper, which allows other modules to themselves create a PolyMapDesugarHelper.
- Do not use "knownFolded" terminology and instead use "predicate mask" terminology to be consistent with other code
