+++ src/main/scala/viper/carbon/modules/impls/DefaultHeapModule.scala
@@ -13,7 +13,7 @@ import viper.silver.{ast => sil}
 import viper.carbon.boogie._
 import viper.carbon.boogie.Implicits._
 import viper.carbon.verifier.Verifier
-import viper.carbon.utility.{MapDesugarHelper, MapRep}
+import viper.carbon.utility.{PolyMapDesugarHelper, PolyMapRep}
 import viper.silver.ast.utility.QuantifiedPermissions.QuantifiedPermissionAssertion
 import viper.silver.verifier.PartialVerificationError
 
@@ -103,10 +103,7 @@ class DefaultHeapModule(val verifier: Verifier)
 
   override def refType = NamedType("Ref")
 
-  override def heapMapDesugarHelper = MapDesugarHelper(
-    refType,
-    (t1: Type, t2: Type) => NamedType(fieldTypeName, Seq(t1, t2)).asInstanceOf[Type],
-    heapNamespace)
+  override def fieldTypeConstructor = (2, (ts: Seq[Type]) => NamedType(fieldTypeName, ts).asInstanceOf[Type])
 
   override def preamble = {
     val obj = LocalVarDecl(Identifier("o")(axiomNamespace), refType)
@@ -118,7 +115,9 @@ class DefaultHeapModule(val verifier: Verifier)
     val predField = LocalVarDecl(Identifier("pm_f")(axiomNamespace),
       predicateVersionFieldType("C"))
     val useSumOfStatesAxioms = loopModule.sumOfStatesAxiomRequired
-    val heapDesugaringRep : MapRep = heapMapDesugarHelper.desugarMap(heapTyp, (readHeapName, updateHeapName), t1 => t1.freeTypeVars(1))
+
+    val heapMapDesugarHelper = PolyMapDesugarHelper(refType, fieldTypeConstructor, heapNamespace)
+    val heapDesugaringRep : PolyMapRep = heapMapDesugarHelper.desugarPolyMap(heapTyp, (readHeapName, updateHeapName), t1 => t1.freeTypeVars(1))
 
     TypeDecl(refType) ++
       GlobalVarDecl(heapName, heapTyp) ++
@@ -503,6 +502,10 @@ class DefaultHeapModule(val verifier: Verifier)
   }
 
 
+  /**
+    * @param maskField the field with which the predicate mask is accessed in the heap
+    * @param mask the predicate mask itself (for example Heap[null, [[maskField]]])
+    */
   case class PredicateMask(maskField: Exp, mask: Exp)
 
   private def predicateMask(loc: sil.PredicateAccess) : PredicateMask = {
@@ -559,7 +562,7 @@ class DefaultHeapModule(val verifier: Verifier)
           @{code FuncApp} are ignored in the default case; the return type is not required for Boogie's type checker in
           general).
        */
-      FuncApp( if(isPMask) { permModule.knownFoldedMaskRep.selectId } else { readHeapName }, Seq(h,o,f), Bool)
+      FuncApp( if(isPMask) { permModule.pmaskTypeDesugared.selectId } else { readHeapName }, Seq(h,o,f), Bool)
     }
   }
 
@@ -587,7 +590,7 @@ class DefaultHeapModule(val verifier: Verifier)
     if(verifier.usePolyMapsInEncoding)
       MapUpdate(heap, Seq(rcv, field), newVal)
     else
-      FuncApp(if(isPMask) { permModule.knownFoldedMaskRep.storeId } else { updateHeapName }, Seq(heap, rcv, field, newVal), Bool)
+      FuncApp(if(isPMask) { permModule.pmaskTypeDesugared.storeId } else { updateHeapName }, Seq(heap, rcv, field, newVal), Bool)
   }
 
   override def translateLocationAccess(f: sil.LocationAccess): Exp = {
