+++ src/main/scala/viper/carbon/modules/impls/QuantifiedPermModule.scala
@@ -36,6 +36,7 @@ import viper.carbon.boogie.Assign
 import viper.carbon.boogie.Func
 import viper.carbon.boogie.TypeAlias
 import viper.carbon.boogie.FuncApp
+import viper.carbon.utility.PolyMapDesugarHelper
 import viper.carbon.verifier.Verifier
 import viper.silver.ast.utility.rewriter.Traverse
 import viper.silver.ast.Implies
@@ -127,15 +128,17 @@ class QuantifiedPermModule(val verifier: Verifier)
   private val readPMaskName = Identifier("readPMask")
   private val updatePMaskName = Identifier("updPMask")
 
-  override val knownFoldedMaskRep = KnownFoldedMaskRep(NamedType(pmaskTypeName), readPMaskName, updatePMaskName)
+  override val pmaskTypeDesugared = PMaskDesugaredRep(readPMaskName, updatePMaskName)
 
   override def preamble = {
     val obj = LocalVarDecl(Identifier("o")(axiomNamespace), refType)
     val field = LocalVarDecl(Identifier("f")(axiomNamespace), fieldType)
     val permInZeroMask = currentPermission(zeroMask, obj.l, field.l)
     val permInZeroPMask = currentPermission(zeroPMask, obj.l, field.l, true)
-    val maskRep = heapMapDesugarHelper.desugarMap(maskType, (readMaskName, updateMaskName), _ => permType)
-    val pmaskRep = heapMapDesugarHelper.desugarMap(pmaskType, (readPMaskName, updatePMaskName), _ => Bool)
+
+    val maskPolyMapDesugarHelper = PolyMapDesugarHelper(refType, fieldTypeConstructor, namespace)
+    val maskRep = maskPolyMapDesugarHelper.desugarPolyMap(maskType, (readMaskName, updateMaskName), _ => permType)
+    val pmaskRep = maskPolyMapDesugarHelper.desugarPolyMap(pmaskType, (readPMaskName, updatePMaskName), _ => Bool)
 
     // permission type
     TypeAlias(permType, Real) ::
@@ -1370,10 +1373,14 @@ class QuantifiedPermModule(val verifier: Verifier)
     currentPermission(maskExp, rcv, location)
   }
   def currentPermission(mask: Exp, rcv: Exp, location: Exp, isPMask: Boolean = false): Exp = {
-    if(verifier.usePolyMapsInEncoding)
+    if(verifier.usePolyMapsInEncoding) {
       MapSelect(mask, Seq(rcv, location))
-    else
-      FuncApp( if(isPMask) { readPMaskName } else { readMaskName }, Seq(mask, rcv, location), permType)
+    } else {
+      FuncApp( if(isPMask) { readPMaskName } else { readMaskName },
+               Seq(mask, rcv, location),
+               if(isPMask) { Bool } else { permType }
+      )
+    }
   }
 
   override def permissionLookup(la: sil.LocationAccess) : Exp = {
